+++ Assets/Scripts/HexFlags.cs
@@ -6,14 +6,34 @@ public enum HexFlags
 {
 	Empty = 0,
 
-	RoadNE = 0b0000_0001,
-	RoadE  = 0b0000_0010,
-	RoadSE = 0b0000_0100,
-	RoadSW = 0b0000_1000,
-	RoadW  = 0b0001_0000,
-	RoadNW = 0b0010_0000,
+	RoadNE = 0b000001,
+	RoadE  = 0b000010,
+	RoadSE = 0b000100,
+	RoadSW = 0b001000,
+	RoadW  = 0b010000,
+	RoadNW = 0b100000,
 
-	Roads = 0b0011_1111
+	Roads = 0b111111,
+
+	RiverInNE = 0b000001_000000,
+	RiverInE  = 0b000010_000000,
+	RiverInSE = 0b000100_000000,
+	RiverInSW = 0b001000_000000,
+	RiverInW  = 0b010000_000000,
+	RiverInNW = 0b100000_000000,
+
+	RiverIn = 0b111111_000000,
+
+	RiverOutNE = 0b000001_000000_000000,
+	RiverOutE  = 0b000010_000000_000000,
+	RiverOutSE = 0b000100_000000_000000,
+	RiverOutSW = 0b001000_000000_000000,
+	RiverOutW  = 0b010000_000000_000000,
+	RiverOutNW = 0b100000_000000_000000,
+
+	RiverOut = 0b111111_000000_000000,
+
+	River = 0b111111_111111_000000,
 }
 
 public static class HexFlagsExtensions
@@ -27,13 +47,22 @@ public static class HexFlagsExtensions
 	public static bool HasAny (this HexFlags flags, HexFlags mask) => (flags & mask) != 0;
 
 	/// <summary>
+	/// Returns flags with bits of the given mask cleared.
+	/// </summary>
+	/// <param name="flags"><Flags./param>
+	/// <param name="mask">Mask to clear.</param>
+	/// <returns>The cleared flags.</returns>
+	public static HexFlags Without (this HexFlags flags, HexFlags mask) =>
+		flags &= ~mask;
+
+	/// <summary>
 	/// Whether the flag for a road in a given direction is set.
 	/// </summary>
 	/// <param name="flags">Flags.</param>
 	/// <param name="direction">Road direction.</param>
 	/// <returns>Whether the road is set.</returns>
 	public static bool HasRoad (this HexFlags flags, HexDirection direction) =>
-		((int)flags & (1 << (int)direction)) != 0;
+		flags.Has(HexFlags.RoadNE, direction);
 
 	/// <summary>
 	/// Returns the flags with the bit for a given road set.
@@ -42,7 +71,7 @@ public static class HexFlagsExtensions
 	/// <param name="direction">Road direction.</param>
 	/// <returns>Flags with the road bit set.</returns>
 	public static HexFlags WithRoad (this HexFlags flags, HexDirection direction) =>
-		flags | (HexFlags)(1 << (int)direction);
+		flags.With(HexFlags.RoadNE, direction);
 
 	/// <summary>
 	/// Returns the flags without the bit for a given road set.
@@ -51,5 +80,99 @@ public static class HexFlagsExtensions
 	/// <param name="direction">Road direction.</param>
 	/// <returns>Flags without the road bit set.</returns>
 	public static HexFlags WithoutRoad (this HexFlags flags, HexDirection direction) =>
-		flags & ~(HexFlags)(1 << (int)direction);
+		flags.Without(HexFlags.RoadNE, direction);
+
+	/// <summary>
+	/// Whether the flag for an incoming river in a given direction is set.
+	/// </summary>
+	/// <param name="flags">Flags.</param>
+	/// <param name="direction">Incoming river direction.</param>
+	/// <returns>Whether the river is set.</returns>
+	public static bool HasRiverIn (this HexFlags flags, HexDirection direction) =>
+		flags.Has(HexFlags.RiverInNE, direction);
+
+	/// <summary>
+	/// Returns the flags with the bit for a given incoming river set.
+	/// </summary>
+	/// <param name="flags">Flags.</param>
+	/// <param name="direction">Incoming river direction.</param>
+	/// <returns>Flags with the river bit set.</returns>
+	public static HexFlags WithRiverIn (this HexFlags flags, HexDirection direction) =>
+		flags.With(HexFlags.RiverInNE, direction);
+
+	/// <summary>
+	/// Returns the flags without the bit for a given incoming river set.
+	/// </summary>
+	/// <param name="flags">Flags.</param>
+	/// <param name="direction">Incoming river direction.</param>
+	/// <returns>Flags without the river bit set.</returns>
+	public static HexFlags WithoutRiverIn (this HexFlags flags, HexDirection direction) =>
+		flags.Without(HexFlags.RiverInNE, direction);
+
+	/// <summary>
+	/// Whether the flag for an outgoing river in a given direction is set.
+	/// </summary>
+	/// <param name="flags">Flags.</param>
+	/// <param name="direction">Outgoing river direction.</param>
+	/// <returns>Whether the river is set.</returns>
+	public static bool HasRiverOut (this HexFlags flags, HexDirection direction) =>
+		flags.Has(HexFlags.RiverOutNE, direction);
+
+	/// <summary>
+	/// Returns the flags with the bit for a given outgoing river set.
+	/// </summary>
+	/// <param name="flags">Flags.</param>
+	/// <param name="direction">Outgoing river direction.</param>
+	/// <returns>Flags with the river bit set.</returns>
+	public static HexFlags WithRiverOut (this HexFlags flags, HexDirection direction) =>
+		flags.With(HexFlags.RiverOutNE, direction);
+
+	/// <summary>
+	/// Returns the flags without the bit for a given outgoing river set.
+	/// </summary>
+	/// <param name="flags">Flags.</param>
+	/// <param name="direction">Outgoing river direction.</param>
+	/// <returns>Flags without the river bit set.</returns>
+	public static HexFlags WithoutRiverOut (
+		this HexFlags flags, HexDirection direction
+	) =>
+		flags.Without(HexFlags.RiverOutNE, direction);
+
+	/// <summary>
+	/// Returns the incoming river direction. Only valid if the river exists.
+	/// </summary>
+	/// <param name="flags">Flags.</param>
+	/// <returns>River direction.</returns>
+	public static HexDirection RiverInDirection (this HexFlags flags) =>
+		flags.ToDirection(6);
+
+	/// <summary>
+	/// Returns the outgoing river direction. Only valid if the river exists.
+	/// </summary>
+	/// <param name="flags">Flags.</param>
+	/// <returns>River direction.</returns>
+	public static HexDirection RiverOutDirection (this HexFlags flags) =>
+		flags.ToDirection(12);
+
+	static bool Has (this HexFlags flags, HexFlags start, HexDirection direction) =>
+		((int)flags & ((int)start << (int)direction)) != 0;
+
+	static HexFlags With (this HexFlags flags, HexFlags start, HexDirection direction) =>
+		flags | (HexFlags)((int)start << (int)direction);
+
+	static HexFlags Without (
+		this HexFlags flags, HexFlags start, HexDirection direction
+	) =>
+		flags & ~(HexFlags)((int)start << (int)direction);
+
+	static HexDirection ToDirection (this HexFlags flags, int shift) =>
+		 (((int)flags >> shift) & 0b111111) switch
+		 {
+			 0b000001 => HexDirection.NE,
+			 0b000010 => HexDirection.E,
+			 0b000100 => HexDirection.SE,
+			 0b001000 => HexDirection.SW,
+			 0b010000 => HexDirection.W,
+			 _ => HexDirection.NW
+		 };
 }